requirements.txt
flask==3.0.3
sqlalchemy==2.0.32
flask-sqlalchemy==3.1.1
psycopg[binary]==3.2.1
python-dotenv==1.0.1
gunicorn==22.0.0
